requirements/base.txt
Django>=2.1,<2.2
pyldap>=2.4,<2.5
django-auth-ldap==1.7.0
djangorestframework>=3.1,<4
django-filter>=2.0,<2.1
django-datetime-widget==0.9.3
boto>=2,<3
paramiko>=2.1,<2.2
django-bootstrap3>=11.0,<11.1
git+https://github.com/justinmayer/django-autoslug@master#egg=autoslug
pyapi-gitlab==7.8.5
django-markupfield==1.4.1
Pillow>=3,<4
git+https://github.com/justquick/django-activity-stream@master#egg=actstream
django-braces>=1.13,<1.14
# django-wysiwyg>=0.8,<0.9
django-ckeditor>=5.1,<5.2
html2text==2016.9.19
django-grappelli>=2.12,<2.13
django-mptt>=0.9,<0.10
git+https://github.com/wbg-optronix-lab/aws-support@0.11.0#egg=aws_support
django-storages==1.7.1
django-ace>=1.0,<1.1
django_polymorphic>=2,<2.1
redis>=3,<4
numpy>=1.15
astropy>=2,<3
git+https://github.com/nnewman/nanoscope@master#egg=nanoscope
django-sendfile==0.3.11
django-betterforms>=1.2,<1.3
ExifRead==2.1.2
django-rq>=1.1.0,<1.2
django-crispy-forms>=1.7,<1.8
mendeley==0.3.2
django-simple-history>=2.6,<2.7
django-jsonfield
django-jsonfield-compat
gunicorn==19.6.0
